refactor(bump results): bump max results to 1000
Putting in proper pagination would mean me learning angular, and that'll take longer than just bumping the max results to 1000. the amount of traffic this project sees, coupled with the fact that there are less than 100k documents in the index in total means bumping the number should remove the need for pagination for most meaningful queries.